Fixed placement of navigation menu
Fixes #5

The navigation bar and the article need to be put in a common <div> tag
for the current style.css to work. It *probably* needs to be in that
<div> tag anyway if we want a proper grid with the title occupying the
entire top row.

Also widened the body and narrowed the navigation menu.
Those widths will almost certainly need further adjustments.