Ruthless nitpicking
Only the landing page and the glossary so far.
Changes include:

- Using === and --- to denote <h1> and <h2> headings.
- Use <h1> for the title only, <h2> and higher for subsections.
- Links that exceed the line length use footnote style references.
- Lots and tons of personnal nitpicks and little rewrites that
  hardly even matter.

The changes look massive, but it's mostly small stuff.
I liked the original text, so I mostly kept it,
even if Git doesn't see it that way.

Also, because this is all versionned, I've tried my best to keep line
breaks after periods and most commas.

TODO:
- Spell checking
- map.txt (tha one is being a bear)