Improve styling of <aside>
The padding is now uniform inside the box and the margin no longer makes
it visually right-leaning.

Font size has been decreased to visually indicate that this aside is
less important than the surrounding text.
To the same effect, the main font color has been made somewhat lighter;
care has been taken not to violate WCAG AAA with the contrast ratio
(which is currently 14:1).

The background color also contains a tiny hint of orange to match with
the rest of the orange-ish highlights on the page.