map: Remove references to faulty implementations
The main faulty implementation with a citeable reference as to why
selecting from the whole curve instead of only the prime-order subgroup
was Adam Langley (agl)'s with no good impartial, third-party page
summarizing multiple such cases
(of which there aren't enough in the first place due to the lack of
implementations -- which is the whole root issue we're trying to
address with this website in the first place).
However, this would unfairly single out agl,
so it ends up being more practical just removing this section entirely
to avoid stepping on any toes.

See discussion on 54a29df2 for additional context.