Removed padding from test vector generation
Padding can be handled in way too many ways:

- The sign bit may be the LSB or the MSB.
- We can pad the sign bit *or* we can negate the representative.
- We may have additional bits to make a multiple of 8 (one byte).

This makes things confusing, and it's not clear how to handle this for
each situation.  For instance, it's probably better for Curve25519 to
pad the sign bit because the sign bit is likely to be the MSB, and we
need to pad an additional bit to complete the last byte.

On the other hand, Curve448 might want to negate the representative
instead, so it can avoid padding altogether, and recipients don't have
to mask any bit.

---

Also, fixed a bug in the test25519.c that caused the tests to fail when
the first test vector is not zero.