Curve448 low order point has negative coordinates
For Curve448 we decided that odd field elements are negative.  Thus, 1
is negative, and -1 is positive, which is quite counter-intuitive.

Though I believe we should still chose (1, 0) as the point of order 4
instead of (-1, 0), we should not pretend it has positive coordinates.
Besides, which we chose doesn't matter that much, it will still be
compatible with an implementation that chose differently.